refactor(Education, Experience): moving utils to correct place, removing unused files and utils
