--- src/modules/Education/Education.constants.ts
@@ -1,21 +1,0 @@
-export const NORMALISED_EDUCATION_MOCK = {
-  ids: ['id1', 'id2'],
-  entities: {
-    id1: {
-      id: 1,
-      metadata: ['BA Degree'],
-      iconUrl: '',
-      title: 'Rhodes University',
-      description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ',
-      isValidated: true,
-    },
-    id2: {
-      id: 2,
-      metadata: ['Matric'],
-      iconUrl: '',
-      title: 'South Africa',
-      description: '',
-      isValidated: false,
-    },
-  },
-}
