--- src/modules/Education/Education.types.ts
@@ -1,8 +1,0 @@
-export type EducationEntry = {
-  qualification: string
-  organisationLogoURL: string
-  school: string
-  startDate: string
-  endDate: string
-  description: string
-}
