+++ src/modules/Education/View/EducationView.util.ts
@@ -1,7 +1,7 @@
 import { applySpec, compose, isEmpty, pipe, propOr, reject, unless, values } from 'ramda'
 
-import { DATE_TPL_MON_YEAR } from '../../constants/date.constants'
-import { formatDateString } from '../../utils/dates.utils'
+import { DATE_TPL_MON_YEAR } from '../../../constants/date.constants'
+import { formatDateString } from '../../../utils/dates.utils'
 
 export const getEducationMetadata = pipe(
   applySpec({
