+++ src/modules/Experience/View/ExperienceView.utils.test.ts
@@ -1,0 +1,19 @@
+import { mergeDeepRight } from 'ramda'
+
+import { USER_JOBS_MOCK } from '../../UserJobs/UserJobs.test.fixtures'
+import * as SUT from './ExperienceView.utils'
+
+describe('modules/Experience/Experience.utils', () => {
+  describe('getExperienceMetadata', () => {
+    it('should return the required metadata from a User Job Credential', () => {
+      const result = SUT.getExperienceMetadata(USER_JOBS_MOCK[0])
+
+      expect(result).toStrictEqual(['NAME', 'Jun 2021 - Aug 2021 • 2 months', 'COUNTRY'])
+    })
+    it('should correctly handle if a property is empty', () => {
+      const result = SUT.getExperienceMetadata(mergeDeepRight(USER_JOBS_MOCK[0], { job: { countries: undefined } }))
+      expect(result.length).toBe(2)
+      expect(result).toStrictEqual(['NAME', 'Jun 2021 - Aug 2021 • 2 months'])
+    })
+  })
+})
